% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawRmap.R
\name{drawRmap}
\alias{drawRmap}
\title{Draw a Map with a Specified Location}
\usage{
drawRmap(location, maptype = "stamen_toner", zoom = 12)
}
\arguments{
\item{location}{A character string representing the location to be plotted.}

\item{maptype}{A character string specifying the type of map to be used.
Valid options are "stamen_toner", "stamen_toner_lite", "stamen_watercolor",
"stamen_terrain", "stamen_terrain_background", "stamen_terrain_labels",
"stamen_terrain_lines", "outdoors", "alidade_smooth", and "alidade_smooth_dark".
Defaults to "stamen_toner".}

\item{zoom}{An integer representing the zoom level of the map, ranging from 1 to 20.
Defaults to 12.}
}
\value{
A ggplot object displaying the map with the specified location marked.
}
\description{
This function retrieves the geographical coordinates for a given location
and plots a map using the specified map type and zoom level. The function
utilizes the Stadia Maps service for rendering the map. A valid API key
is required for accessing the Stadia Maps service.
}
\examples{
\dontrun{
# Draw map for Asia
drawRmap("Asia", maptype = "stamen_toner", zoom = 10)

# Draw map for Harvard University
drawRmap("Harvard University", maptype = "outdoors", zoom = 11)
}
}
